chore: refactor operators file
