+++ client/app/(authenticated)/idir/cas_admin/dashboard/operators/page.tsx
@@ -1,4 +1,4 @@
-import OperatorsPage from "@/app/components/routes/operators/OperatorsPage";
+import OperatorsPage from "@/app/components/routes/operators/Page";
 import { OperatorsSearchParams } from "@/app/components/userOperators/types";
 
 export default async function Page({
